Add credits to About box
They're a bit bland right now, and not even big enough to create a
scrolling text field, so the About box actually looks kind of stupid.
I tried to fix that by adding a blank space at the top, but that
made the window render with the text scroll view partially scrolled
down (not sure why, although I feel like I've had this problem
before). At any rate, I'm sure I'll add more info later on,
anyway.

The style is inspired by that of TextMate's About box; not sure if
I'm going to keep this style or not.