Only compile for 64-bit Intel
The plan is to convert Blackout into a completely faceless app that is
invoked with a global key combo (to save menubar space and just make
the app sexier in general). Preferences will be controlled via a
preference pane. 32-bit prefpanes are annoying (and semi-deprecated)
on Mac OS X 10.6. Is anyone on 10.6 still using 32-bit Intel
processors, anyway?