Make Blackout.app a faceless background app
Blackout.app will be a "helper" app that will run in the background
and listen for the appropriate key events for activating the
screen saver. A preference pane will control options for Blackout;
Blackout itself will not have a menu or any screen presence at all.
This is sexier than using up valuable status menu item space for an
app that can easily be invoked via the keyboard.