Bump the build number
There hasn't been any major changes, but I want to test to see if
the Sparkle updating mechanism is working. I've never actually
used Sparkle to update a preference pane bundle before; I want to
see if the code is actually correct.

Thankfully, it works; however, the preference pane does not
automatically get reloaded, and so shows the old version until the
user quits System Preferences and reloads the pane. I'll have to
look into how to fix this.