Alert user when no Sparkle updates are available
The Sparkle dialog removes focus from System Preferences, so after
clicking OK, the System Preferences window is no longer in focus. I
don't like this, but I don't see a good way to prevent this behavior,
either. I'd like to use an alert sheet to inform the user, but Sparkle
does not send its delegate messages with its update status when the
update is checked in the background, so I cannot do that. This will
have to suffice for now.