Use 10.6-only functionality on 10.6
The helper and prefpane now weaklink against the 10.6 SDK. On 10.5
it will continue to use the older, Carbon-y APIs for getting info
and notifications about running apps; on 10.6, it will use the shiny
new 10.6 APIs.

I've never used weaklinking before, so I'm a bit nervous about this
change -- it could make testing a bit more difficult -- but I think
it's worth it to be a bit future-proof.