Remove prefpane and build as a daemon
Managing the interaction between the prefpane and the daemon was
becoming cumbersome. I think having a simple daemon is best. My idea
now is to have the preferences window activated by, e.g., holding in
Option when opening the app. Perhaps I'll also show the window the
first time the app is launched (with an explanation as to how to show
it again). Maybe I'll even prompt to add to the login items the first
time the app is show. At any rate, this will be a lot simpler, and if
I want to get "fancier", I can pull out some of the old code I was using
in the prefpane for managing options.