Response to Command-W keystrokes
As noted in [1], the easiest way to have a macOS background app respond
to keyboard events is to implement a main menu. One way to do that is to
copy the main menu from the default Xcode Cocoa Application template.
Another way is to create a menu that contains submenus, each with an
item you want to implement (like "Close"). The problem is that there is
not an easy way in the Xcode interface to designate a menu as the "main"
menu. What you can do is create a menu, then open up the MainMenu.xib
file in a text editor like vim. Find the XML block corresponding to the
menu you just created, and add a `systemMenu="main"` attribute. When you
open your MainMenu.xib file in Xcode again, the menu should look like a
main menu, instead of the usual vertical menu.

  [1] https://hsoienterprises.com/2014/07/24/lsuielement-and-cmd-w/