Upgrade project for Xcode 14.2
I'm amazed that after a decade this project still compiles! It needed a
few updates to project settings to silence warnings in the latest
version of Xcode. It also can't be built on arm64 at all, since the
ShortcutRecorder framework only has code for x86_64; I'll either have to
recompile a new version of that framework or only compile Blackout for
x86_64.