Support arm64 in ShortcutRecorder.framework
I had to compile a new version of this framework from source since it
hasn't been touched in about a decade, but this version includes an
arm64 build and _should_ work fine. It is obviously a newer version than
the one that I was previously using; it ditches support for Leopard (but
Blackout doesn't support Leopard anyway -- who does anymore?); and it
includes more localizations.