Deploy to macOS 10.13
I'm really curious where the $(RECOMMENDED_MACOSX_DEPLOYMENT_TARGET)
variable came from, because I can't find that documented _anywhere_.
Maybe it's from an earlier version of Xcode; at any rate, it was causing
glitches in the UI wherever the deployment target was shown in Xcode, so
let's just set it to the minimum available right now, 10.13, and leave
it at that.

Of course, at this moment, the recommended minimum deployment target is
macOS 10.14.6, which was released in 2018, so to be honest, it wouldn't
necessarily be a bad idea to just set that as the minimum.