Enable Apple Generic Versioning
I have _obsessed_ over the CFBundleVersion ever since I started writing
Mac software almost 20 years ago, but I have come to the conclusion that
the best bet is to just use a monotonically-increasing integer and be
done with it -- screw trying to convey that a build is the nth build
_for a given version_. When do I really have to worry about bug fixes
vs. major version numbers anyway? Besides, maybe I'll start using CalVer
for my marketing version and be done with it.

I have also enabled Apple Generic Versioning which can automatically
increment the build number on every build. I have elected to only
increment the build number on _releases_ for the time being, but I can
re-evaluate in the future if I want to increment the version number on
every single build, including test runs.